requirements.txt
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
PyYAML>=6.0
reportlab>=4.0
